correction boucle infinie quand GPS branché pendant une fonction bloquante (#8578)
git-svn-id: http://einstein/svn/firmware/FIR-00251-Proc/trunk@19728 3e98269d-ff51-e449-abca-491a6567d830

+++ src/FW/Driver/FaulhaberProtocol.c
@@ -242,7 +242,7 @@ void FH_parseResponse(FH_ctrl_t* chan)
             case FHR_STRING:
             {
                char tmp[64];
-               uint8_t length = CBB_Length(partialResp);
+               uint8_t length = (uint8_t)CBB_Length(partialResp);
 
                tmp[length] = 0;
                CBB_Popn(partialResp, CBB_Length(partialResp), (uint8_t*)tmp);
@@ -606,7 +606,7 @@ uint8_t FH_readAcks(FH_ctrl_t* data)
       --data->fh_data.numResponses;
       ++numRead;
 
-      length = CBB_Length(&data->fh_data.respTypes);
+      length = (uint8_t)CBB_Length(&data->fh_data.respTypes);
       status = CBB_Peek(&data->fh_data.respTypes, 0, &respType);
    }
